backend/requirements.txt
flask
pillow
rembg
onnxruntime